Delete ColourRoleDBManager.py. Move command check for colour change to RoleColourCog.py. Remove tests for check from test_KoalaBot.py
