--- utils/ColourRoleDBManager.py
@@ -1,74 +1,0 @@
-#!/usr/bin/env python
-
-"""
-KoalaBot Cog for guild members wishing to change their role colour
-"""
-
-# Futures
-
-# Built-in/Generic Imports
-
-# Libs
-
-# Own modules
-import KoalaBot
-from utils import KoalaDBManager
-
-# Constants
-
-# Variables
-
-
-class ColourRoleDBManager:
-    """
-    A class for interacting with the Koala Colour Role database
-    """
-
-    def __init__(self, database_manager: KoalaDBManager):
-        self.database_manager = database_manager
-
-    def get_parent_database_manager(self):
-        return self.database_manager
-
-    def create_tables(self):
-        """
-        Creates all the tables associated with the Custom Colour Role extension
-        """
-        # GuildColourChangePermissions
-        sql_create_guild_colour_change_permissions_table = """
-        CREATE TABLE IF NOT EXISTS GuildColourChangePermissions (
-        guild_id integer NOT NULL,
-        role_id integer NOT NULL,
-        PRIMARY KEY (guild_id, role_id),
-        FOREIGN KEY (guild_id) REFERENCES GuildExtensions (guild_id)
-        );"""
-
-        # GuildInvalidCustomColours
-        sql_create_guild_colour_change_invalid_colours_table = """
-        CREATE TABLE IF NOT EXISTS GuildInvalidCustomColourRoles (
-        guild_id integer NOT NULL,
-        role_id integer NOT NULL,
-        PRIMARY KEY (guild_id, role_id),
-        FOREIGN KEY (guild_id) REFERENCES GuildExtensions (guild_id)
-        );"""
-
-        # Create Tables
-        self.database_manager.db_execute_commit(sql_create_guild_colour_change_permissions_table)
-        self.database_manager.db_execute_commit(sql_create_guild_colour_change_invalid_colours_table)
-
-    def add_colour_change_role_perms(self, guild_id, role_id):
-        self.database_manager.db_execute_commit(
-            f"""INSERT INTO GuildColourChangePermissions (guild_id, role_id) VALUES ({guild_id}, {role_id});""")
-
-    def remove_colour_change_role_perms(self, guild_id, role_id):
-        self.database_manager.db_execute_commit(
-            f"""DELETE FROM GuildColourChangePermissions WHERE guild_id = {guild_id} AND role_id = {role_id};""")
-
-    def add_guild_protected_colour_role(self, guild_id, role_id):
-        self.database_manager.db_execute_commit(
-            f"""INSERT INTO GuildInvalidCustomColourRoles (guild_id, role_id) VALUES ({guild_id}, {role_id});""")
-
-    def remove_guild_protected_colour_role(self, guild_id, role_id):
-        self.database_manager.db_execute_commit(
-            f"""DELETE FROM GuildInvalidCustomColourRoles WHERE guild_id = {guild_id} AND role_id = {role_id};""")
-
